requirements.txt
pandas
numpy
scipy
scikit-learn
tqdm
networkx
matplotlib
umap-learn
bnlearn
tabulate
decoupler